Adding unit test for saving task

+++ backend/modules/task_management/service/task_service.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"errors"
 	taskManagementContract "stm/modules/task_management/domain/contract"
 	taskDto "stm/modules/task_management/domain/dto"
 	taskManagementEntity "stm/modules/task_management/domain/entity"
@@ -8,16 +9,21 @@ import (
 )
 
 type task struct {
-	taskRepo           taskManagementContract.TaskRepository
-	taskManagementRepo taskManagementContract.TaskRepository
+	taskRepo taskManagementContract.TaskRepository
 }
 
 func (l *task) Save(task *taskManagementEntity.Task, userId string) (taskDto.TaskRecord, error) {
 	// Setting logged user as the creator of the task
 	if task.Id == 0 {
 		task.CreatedById = userId
+	} else {
+		// Prevent that another user update the task if it's not the owner
+		if task.CreatedById != userId {
+			return taskDto.TaskRecord{}, errors.New("you are not allowed to edit this task")
+		}
 	}
 
+	// Validating task information
 	if err := task.Validate(); err != nil {
 		return taskDto.TaskRecord{}, err
 	}
@@ -28,7 +34,7 @@ func (l *task) Save(task *taskManagementEntity.Task, userId string) (taskDto.Tas
 		return taskDto.TaskRecord{}, err
 	}
 
-	return l.taskManagementRepo.GetTaskRecordById(taskRecord.Id)
+	return l.taskRepo.GetTaskRecordById(taskRecord.Id)
 }
 
 func (l *task) GetAllTasks(filter *sharedModel.CriteriaFilter) ([]taskDto.TaskRecord, int64, error) {
@@ -36,6 +42,10 @@ func (l *task) GetAllTasks(filter *sharedModel.CriteriaFilter) ([]taskDto.TaskRe
 }
 
 func (l *task) GetTaskById(taskId uint) (taskManagementEntity.Task, error) {
+	if taskId == 0 {
+		return taskManagementEntity.Task{}, errors.New("not allowed param")
+	}
+
 	task, err := l.taskRepo.GetTaskById(taskId)
 
 	if err != nil {
@@ -46,6 +56,10 @@ func (l *task) GetTaskById(taskId uint) (taskManagementEntity.Task, error) {
 }
 
 func (l *task) ChangeTaskStatus(taskId uint, status uint) error {
+	if taskId == 0 {
+		return errors.New("not allowed param")
+	}
+
 	return l.taskRepo.ChangeTaskStatus(taskId, status)
 }
 
@@ -54,15 +68,17 @@ func (l *task) GetAllTasksStatus() ([]taskManagementEntity.TaskStatus, error) {
 }
 
 func (l *task) DeleteTask(taskId uint) error {
+	if taskId == 0 {
+		return errors.New("not allowed param")
+	}
+
 	return l.taskRepo.DeleteTask(taskId)
 }
 
 func NewTaskService(
 	taskRepo taskManagementContract.TaskRepository,
-	taskManagementRepo taskManagementContract.TaskRepository,
 ) taskManagementContract.TaskService {
 	return &task{
-		taskRepo:           taskRepo,
-		taskManagementRepo: taskManagementRepo,
+		taskRepo: taskRepo,
 	}
 }
